Some hotfixes.

+++ aygp_backend_sql/src/main/java/com/nicogbdev/aygp_backend_sql/suceso_clave/application/mapper/dto/SucesoClaveDTO.java
@@ -13,7 +13,7 @@ public class SucesoClaveDTO {
     private Date fechaSuceso;
     private String titulo;
     private String contenido;
-    private Integer valoracion;
+    private Double valoracion;
 
     // Constructor por defecto.
     public SucesoClaveDTO() {
@@ -61,11 +61,11 @@ public class SucesoClaveDTO {
         this.contenido = contenido;
     }
 
-    public Integer getValoracion() {
+    public Double getValoracion() {
         return valoracion;
     }
 
-    public void setValoracion(Integer valoracion) {
+    public void setValoracion(Double valoracion) {
         this.valoracion = valoracion;
     }
